backend/requirements.txt
asyncpg>=0.29.0
redis>=5.0.0
aiohttp>=3.9.0
ccxt>=4.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
